Relocate reward_wrappers to rewards and rename tabular_irl->mce_irl (#356)
* Move reward_wrapper to rewards

* Rename tabular_irl to mce_irl
