+++ docs/modules/imitation.rewards.rst
@@ -9,6 +9,11 @@ imitation.rewards.reward\_nets module
 
 .. automodule:: imitation.rewards.reward_nets
 
+imitation.rewards.reward\_wrapper module
+----------------------------------------
+
+.. automodule:: imitation.rewards.reward_wrapper
+
 imitation.rewards.serialize
 ---------------------------
